[AutoMM] FT_Transformer refactor (#3621)
