+++ examples/automm/tabular_dl/example_tabular.py
@@ -35,8 +35,8 @@ TABULAR_DATASETS = {
 
 automm_hyperparameters = {
     "data.categorical.convert_to_text": False,
-    "model.names": ["categorical_transformer", "numerical_transformer", "fusion_transformer"],
-    "model.numerical_transformer.embedding_arch": ["linear"],
+    "model.names": ["ft_transformer"],
+    "model.ft_transformer.embedding_arch": ["linear"],
     "env.batch_size": 128,
     "env.per_gpu_batch_size": 128,
     "env.eval_batch_size_ratio": 1,
