+++ multimodal/src/autogluon/multimodal/configs/model/fusion_mlp_image_text_tabular.yaml
@@ -8,22 +8,6 @@ model:
     normalization: "layer_norm"
     data_types:
       - "categorical"
-  categorical_transformer:
-    out_features: 192
-    d_token: 192
-    ffn_d_hidden: 192 
-    num_trans_blocks: 0
-    num_attn_heads: 8
-    residual_dropout: 0.0
-    attention_dropout: 0.2
-    ffn_dropout: 0.1
-    normalization: layer_norm
-    ffn_activation: reglu
-    head_activation: relu
-    data_types:
-    - categorical
-    additive_attention: False # Whether to use lightweight additive attention, can be True, False or "auto"
-    share_qv_weights: False # Whether to share weight for query and value, can be True, False or "auto"
   numerical_mlp:
     hidden_size: 128
     activation: "leaky_relu"
@@ -35,26 +19,6 @@ model:
     data_types:
       - "numerical"
     merge: "concat"
-  numerical_transformer:
-    out_features: 192
-    d_token: 192
-    ffn_d_hidden: 192 
-    num_trans_blocks: 0
-    num_attn_heads: 8
-    residual_dropout: 0.0
-    attention_dropout: 0.2
-    ffn_dropout: 0.1
-    normalization: "layer_norm"
-    ffn_activation: "reglu"
-    head_activation: "relu"
-    data_types:
-      - "numerical"
-    embedding_arch:
-      - 'linear'
-      - 'relu'
-    merge: "concat"
-    additive_attention: False # Whether to use lightweight additive attention, can be True, False or "auto"
-    share_qv_weights: False # Whether to share weight for query and value, can be True, False or "auto"
   hf_text:
     checkpoint_name: "google/electra-base-discriminator"
     gradient_checkpointing: False
@@ -251,3 +215,26 @@ model:
     data_types:
     additive_attention: False # Whether to use lightweight additive attention, can be True, False or "auto"
     share_qv_weights: False # Whether to share weight for query and value, can be True, False or "auto"
+  ft_transformer:
+    data_types:
+      - "categorical"
+      - "numerical"
+    embedding_arch:
+      - 'linear'
+    d_token: 192
+    adapter_output_feature: 192
+    hidden_size: 192
+    n_blocks: 3
+    attention_n_heads: 8
+    attention_dropout: 0.2
+    residual_dropout: 0.0
+    ffn_dropout: 0.1
+    ffn_d_hidden: 192
+    ffn_activation: "geglu"
+    head_activation: "relu"
+    normalization: "layer_norm"
+    merge: "concat"
+    requires_all_dtypes: False
+    additive_attention: False # Whether to use lightweight additive attention, can be True, False or "auto"
+    share_qv_weights: False # Whether to share weight for query and value, can be True, False or "auto"
+    pooling_mode: "cls"
