+++ multimodal/src/autogluon/multimodal/constants.py
@@ -231,11 +231,10 @@ HF_TEXT = "hf_text"
 T_FEW = "t_few"
 NUMERICAL_MLP = "numerical_mlp"
 CATEGORICAL_MLP = "categorical_mlp"
-NUMERICAL_TRANSFORMER = "numerical_transformer"
-CATEGORICAL_TRANSFORMER = "categorical_transformer"
 FUSION = "fusion"
 FUSION_MLP = f"{FUSION}_mlp"
 FUSION_TRANSFORMER = f"{FUSION}_transformer"
+FT_TRANSFORMER = "ft_transformer"
 FUSION_NER = f"{FUSION}_{NER}"
 MMDET_IMAGE = "mmdet_image"
 MMOCR_TEXT_DET = "mmocr_text_detection"
