+++ multimodal/src/autogluon/multimodal/utils/model.py
@@ -13,10 +13,10 @@ from ..constants import (
     AUTOMM,
     CATEGORICAL,
     CATEGORICAL_MLP,
-    CATEGORICAL_TRANSFORMER,
     CLIP,
     DOCUMENT,
     DOCUMENT_TRANSFORMER,
+    FT_TRANSFORMER,
     FUSION_MLP,
     FUSION_NER,
     FUSION_TRANSFORMER,
@@ -29,7 +29,6 @@ from ..constants import (
     NER_TEXT,
     NUMERICAL,
     NUMERICAL_MLP,
-    NUMERICAL_TRANSFORMER,
     OVD,
     T_FEW,
     TEXT,
@@ -40,9 +39,9 @@ from ..constants import (
 from ..data import MultiModalFeaturePreprocessor
 from ..models import (
     CategoricalMLP,
-    CategoricalTransformer,
     CLIPForImageText,
     DocumentTransformer,
+    FT_Transformer,
     HFAutoModelForNER,
     HFAutoModelForTextPrediction,
     MMDetAutoModelForObjectDetection,
@@ -52,7 +51,6 @@ from ..models import (
     MultimodalFusionNER,
     MultimodalFusionTransformer,
     NumericalMLP,
-    NumericalTransformer,
     OVDModel,
     TFewModel,
     TimmAutoModelForImagePrediction,
@@ -114,6 +112,7 @@ def select_model(
     fusion_model_name = []
     for model_name in names:
         model_config = getattr(config.model, model_name)
+        strict = getattr(model_config, "requires_all_dtypes", strict)
         if not model_config.data_types:
             fusion_model_name.append(model_name)
             continue
@@ -136,12 +135,7 @@ def select_model(
         assert len(fusion_model_name) == 1
         selected_model_names.extend(fusion_model_name)
     elif len(fusion_model_name) == 1 and hasattr(config.model, fusion_model_name[0]):
-        # TODO: Support using categorical_transformer or numerical_transformer alone without a fusion model.
-        if selected_model_names[0].lower().startswith((CATEGORICAL_TRANSFORMER, NUMERICAL_TRANSFORMER)):
-            selected_model_names.extend(fusion_model_name)
-        else:
-            # remove the fusion model's config make `config.model.names` and the keys of `config.model` consistent.
-            delattr(config.model, fusion_model_name[0])
+        delattr(config.model, fusion_model_name[0])
 
     config.model.names = selected_model_names
     logger.debug(f"selected models: {selected_model_names}")
@@ -245,29 +239,6 @@ def create_model(
             embedding_arch=OmegaConf.select(model_config, "embedding_arch"),
             num_classes=num_classes,
         )
-    elif model_name.lower().startswith(NUMERICAL_TRANSFORMER):
-        model = NumericalTransformer(
-            prefix=model_name,
-            in_features=num_numerical_columns,
-            out_features=model_config.out_features,
-            d_token=model_config.d_token,
-            n_blocks=model_config.num_trans_blocks,
-            attention_n_heads=model_config.num_attn_heads,
-            attention_dropout=model_config.attention_dropout,
-            residual_dropout=model_config.residual_dropout,
-            ffn_dropout=model_config.ffn_dropout,
-            attention_normalization=model_config.normalization,
-            ffn_normalization=model_config.normalization,
-            head_normalization=model_config.normalization,
-            ffn_activation=model_config.ffn_activation,
-            head_activation=model_config.head_activation,
-            cls_token=False,
-            embedding_arch=model_config.embedding_arch,
-            num_classes=num_classes,
-            ffn_d_hidden=OmegaConf.select(model_config, "ffn_d_hidden", default=192),
-            additive_attention=OmegaConf.select(model_config, "additive_attention", default=False),
-            share_qv_weights=OmegaConf.select(model_config, "share_qv_weights", default=False),
-        )
     elif model_name.lower().startswith(CATEGORICAL_MLP):
         model = CategoricalMLP(
             prefix=model_name,
@@ -278,28 +249,6 @@ def create_model(
             dropout_prob=model_config.drop_rate,
             normalization=model_config.normalization,
             num_classes=num_classes,
-        )
-    elif model_name.lower().startswith(CATEGORICAL_TRANSFORMER):
-        model = CategoricalTransformer(
-            prefix=model_name,
-            num_categories=num_categories,
-            out_features=model_config.out_features,
-            d_token=model_config.d_token,
-            n_blocks=model_config.num_trans_blocks,
-            attention_n_heads=model_config.num_attn_heads,
-            attention_dropout=model_config.attention_dropout,
-            residual_dropout=model_config.residual_dropout,
-            ffn_dropout=model_config.ffn_dropout,
-            attention_normalization=model_config.normalization,
-            ffn_normalization=model_config.normalization,
-            head_normalization=model_config.normalization,
-            ffn_activation=model_config.ffn_activation,
-            head_activation=model_config.head_activation,
-            ffn_d_hidden=OmegaConf.select(model_config, "ffn_d_hidden", default=192),
-            num_classes=num_classes,
-            cls_token=False,
-            additive_attention=OmegaConf.select(model_config, "additive_attention", default=False),
-            share_qv_weights=OmegaConf.select(model_config, "share_qv_weights", default=False),
         )
     elif model_name.lower().startswith(DOCUMENT_TRANSFORMER):
         model = DocumentTransformer(
@@ -394,6 +343,31 @@ def create_model(
             additive_attention=OmegaConf.select(model_config, "additive_attention", default=False),
             share_qv_weights=OmegaConf.select(model_config, "share_qv_weights", default=False),
         )
+    elif model_name.lower().startswith(FT_TRANSFORMER):
+        model = FT_Transformer(
+            prefix=model_name,
+            num_numerical_columns=num_numerical_columns,
+            num_categories=num_categories,
+            embedding_arch=model_config.embedding_arch,
+            d_token=model_config.d_token,
+            adapter_output_feature=model_config.adapter_output_feature,
+            hidden_features=model_config.hidden_size,
+            num_classes=num_classes,
+            n_blocks=model_config.n_blocks,
+            attention_n_heads=model_config.attention_n_heads,
+            attention_dropout=model_config.attention_dropout,
+            attention_normalization=model_config.normalization,
+            ffn_d_hidden=model_config.ffn_d_hidden,
+            ffn_dropout=model_config.ffn_dropout,
+            ffn_normalization=model_config.normalization,
+            ffn_activation=model_config.ffn_activation,
+            residual_dropout=model_config.residual_dropout,
+            head_normalization=model_config.normalization,
+            head_activation=model_config.head_activation,
+            additive_attention=OmegaConf.select(model_config, "additive_attention", default=False),
+            share_qv_weights=OmegaConf.select(model_config, "share_qv_weights", default=False),
+            pooling_mode=OmegaConf.select(model_config, "pooling_mode", default="cls"),
+        )
     else:
         raise ValueError(f"unknown model name: {model_name}")
 
@@ -474,11 +448,7 @@ def create_fusion_model(
         # must have one fusion model if there are multiple independent models
         return fusion_model(models=single_models)
     elif len(single_models) == 1:
-        if isinstance(single_models[0], NumericalTransformer) or isinstance(single_models[0], CategoricalTransformer):
-            # TODO: Support using categorical_transformer or numerical_transformer alone without a fusion model.
-            return fusion_model(models=single_models)
-        else:
-            return single_models[0]
+        return single_models[0]
     else:
         raise ValueError(f"No available models for {names}")
